rig/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
pillow>=8.0.0
pygame>=2.0.0
panda3d>=1.10.0
gymnasium>=0.26.0
pyzmq>=22.0.0
opencv-python>=4.5.0
pysdl2>=0.9.0
evdev>=1.4.0
pynput>=1.7.0
psutil>=5.8.0
